+++ MyCrypto.Core/Models/Transaction.cs
@@ -5,11 +5,11 @@ namespace MyCrypto.Core.Models
         public int Id { get; set; }
         public int AddedCoinId { get; set; }
         public int Type { get; set; }
-        public int Price { get; set; } 
-        public int Quantity { get; set; }
-        public int Fee { get; set; }
-        public int Cost { get; set; }
-        public int Earned { get; set; }
+        public double Price { get; set; } 
+        public double Quantity { get; set; }
+        public double Fee { get; set; }
+        public double Cost { get; set; }
+        public double Earned { get; set; }
 
         public virtual AddedCoin AddedCoin { get; set; }
     }
